+++ Sources/AscendKitCore/Support/AscendKitVersion.swift
@@ -1,3 +1,3 @@
 public enum AscendKitVersion {
-    public static let current = "1.8.0"
+    public static let current = "1.9.0"
 }